backend/requirements.txt
fastapi
uvicorn
httpx>=0.24.0
python-dotenv
pydantic
requests
openai[async-extras]
langchain
faiss-cpu
numpy
rich
